requirements.txt
streamlit
google-generativeai
pandas
plotly
PyMuPDF
werkzeug
tabulate
